requirements.txt
web3==6.11.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
jupyter==1.0.0
notebook==7.0.6
